rename API files to match the other
