+++ src/login/components/DataTable/DataTable.tsx
@@ -2,7 +2,7 @@ import React, { Fragment } from "react";
 import EduIDButton from "components/EduIDButton";
 import { FormattedMessage } from "react-intl";
 import { EmailInfo } from "../../../apis/eduidEmail";
-import { PDPhone } from "../../../apis/personalData";
+import { PDPhone } from "../../../apis/eduidPersonalData";
 
 interface DataTableProps {
   data?: EmailInfo[] | PDPhone[];
