+++ src/sagas/Nins.js
@@ -1,4 +1,4 @@
-import { fetchNins } from "apis/personalData";
+import { fetchNins } from "apis/eduidPersonalData";
 import { put } from "redux-saga/effects";
 
 // wrapper from sagas to thunk, remove as soon as possible
